requirements.txt
flet>=0.84.0
flet-video
faster-whisper
ultralytics
librosa
easyocr
ffmpeg-python
numpy<2.0.0
opencv-python
torch --index-url https://download.pytorch.org/whl/cu121
torchvision --index-url https://download.pytorch.org/whl/cu121
torchaudio --index-url https://download.pytorch.org/whl/cu121